elfdefinitions: Suppress unwanted M4 macro expansion.
POSIX does not specify the behavior of the "shift" M4 macro when
the macro is not immediately followed by a left parenthesis.  GNU M4
passes through the macro unexpanded, while NetBSD's M4 expands it to
an empty string.  Adding a level of quotes allows both varieties of
M4 to generate identical content.


git-svn-id: https://svn.code.sf.net/p/elftoolchain/code/trunk@4150 95820547-d848-0410-985e-9ae8fe0fa350