readelf: Replace custom code with the use of an -lelftc API.
Using ELFTC(3) APIs here reduces maintenance overheads when a new
EM_* constant needs to be supported.

This change alters the output of "readelf -h" for a few ELF machine
types.  This should hopefully not be an issue in practice, since the
output of readelf(1) is primarily meant for human consumption.


git-svn-id: https://svn.code.sf.net/p/elftoolchain/code/trunk@4188 95820547-d848-0410-985e-9ae8fe0fa350