libelf/testsuite: handle unsigned enumeration specifiers better.
Per the C standard (section 6.7.2.2 "Enumeration specifiers"),
it is implementation-defined whether the underlying type of an
enumeration is signed or unsigned.  Enumeration constants
themselves are required to be of type 'int' (section 6.4.4.3
"Enumeration constants").

In compilation environments that use an unsigned enumeration
types, an expression like "Elf_Cmd c = ELF_C_NULL - 1" could
end up assigning a large positive unsigned value to "c", which
is not what is desired.

Detected by: Coverity Scan


git-svn-id: https://svn.code.sf.net/p/elftoolchain/code/trunk@4212 95820547-d848-0410-985e-9ae8fe0fa350